% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.MSE.R
\name{get.MSE}
\alias{get.MSE}
\alias{get.MSE.default}
\alias{get.MSE.matrix}
\alias{get.MSE.data.frame}
\title{Gets the MSE between two vectors.}
\usage{
get.MSE(sim, obs, ...)

\method{get.MSE}{default}(sim, obs, ...)

\method{get.MSE}{matrix}(sim, obs, ...)

\method{get.MSE}{data.frame}(sim, obs, ...)
}
\arguments{
\item{sim}{A data frame with the predicted values}

\item{obs}{A data frame with the observed values}

\item{...}{Variables passed to an internal call to the
\code{colMeans()} function}
}
\value{
A measure of the mean squared error between
    the two specified vectors.

A vector with column-wise measures of the
    mean squared error between the specified matrices.

A vector with column-wise measures of the
    mean squared error between the specified data frames
}
\description{
Calculates mean squared error between two vectors.
}
\details{
This is the default implementation that works with
two vectors of classes "integer", "numeric" or "ts".

In this implementation, the arguments are numeric
matrices of the same dimensions.

In this implementation, the arguments are numeric
data frames of the same dimensions.
}
\section{Methods (by class)}{
\itemize{
\item \code{get.MSE(default)}: Default implementation

\item \code{get.MSE(matrix)}: Implementation for matrices

\item \code{get.MSE(data.frame)}: Implementation for data frames

}}
